refactor: extract AudioRecorderCaptureFormat.swift from AudioRecorderTypes.swift
Extracts the `AudioRecorderCaptureFormat` enum (~35 lines, exposes
`makeSettings()` for AVAudioRecorder configuration) into its own file.
AudioRecorderTypes.swift retains the RecordedAudio value, the
AudioRecorderEngine protocol + AVAudioRecorder conformance, and the
factory typealias.

Byte-identical move of lines 23-57. Both files retain their imports;
the new file only needs Foundation.

AudioRecorderTypes.swift: 57 → 21 lines. Tests: 667.
